+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Some MySQL improvements
 * Moved some bad console messages to debugger
 * Some small fixes for errors from Error service
+* Added PlaceholderAPI placeholders support in scoreboard
 
 ### 3.10.1 Release (13/17.08.2018)
 * Fixed NullPointerException in combust event (reported anonymously via Error service) (#1 error service report)